LayoutTests/accessibility/mac/label-element-all-text-string-value.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input type="radio" id="input" />
<label for="input" id="label">
    <span>first</span>
    <span>choice</span>
</label>

<p id="description"></p>
<div id="console"></div>

<script>

    var label = 0;

    description("This tests that if a label element contains text children it's AXValue will be their concatenated AXValues.");
    if (window.accessibilityController) {

        if (window.accessibilityController) {

            var body = document.getElementById("body");
            body.focus();
            
            label = accessibilityController.accessibleElementById("label");
            shouldBe("label.role", "'AXRole: AXStaticText'");
            shouldBe("label.stringValue", "'AXValue: first choice'");          
        }
    }
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
